+++ examples/fortran/dejong.f
@@ -17,7 +17,7 @@ c
 
       common /data/     BinLen, NumCoords, Lower, Upper, gray_on
 
-      integer ctx
+      integer(8)        ctx
       integer           testnum, maxiter, ierror
 
 
@@ -57,7 +57,7 @@ c
 
       double precision function GetTerm(ctx, p, pop, t, problem) 
       include           'pgapackf.h'
-      integer ctx
+      integer(8)         ctx
       integer            p, pop, t, problem, len
       integer            gray_on, BinLen(5), NumCoords(5)
       double precision   Lower(5), Upper(5)
@@ -83,7 +83,7 @@ c
 
       double precision function dejong1(ctx, p, pop) 
       include          'pgapackf.h'
-      integer ctx
+      integer(8)        ctx
       integer           p, pop
       double precision  GetTerm
       external          GetTerm 
@@ -107,7 +107,7 @@ c
 
       double precision function dejong2(ctx, p, pop) 
       include          'pgapackf.h'
-      integer ctx
+      integer(8)        ctx
       integer           p, pop
       double precision  GetTerm
       external          GetTerm
@@ -128,7 +128,7 @@ c
 
       double precision function dejong3(ctx, p, pop) 
       include          'pgapackf.h'
-      integer ctx
+      integer(8)        ctx
       integer           p, pop, i
       double precision  GetTerm
       external          GetTerm
@@ -153,7 +153,7 @@ c
 
       double precision function dejong4(ctx, p, pop) 
       include          'pgapackf.h'
-      integer ctx
+      integer(8)        ctx
       integer           p, pop
       double precision  GetTerm
       external          GetTerm
@@ -177,7 +177,7 @@ c
 
       double precision function dejong5(ctx, p, pop) 
       include          'pgapackf.h'
-      integer ctx
+      integer(8)        ctx
       integer           p, pop, i, j
       double precision  GetTerm
       external          GetTerm
@@ -243,7 +243,7 @@ c
       subroutine printResultInterpretation(ctx, problem)
       include          'pgapackf.h'
       include          'mpif.h'
-      integer ctx
+      integer(8)        ctx
       integer           problem, best, i
       double precision  GetTerm
       external          GetTerm
