+++ examples/fortran/namefull.f
@@ -41,8 +41,8 @@ c     Also, this is common, sunce we need it in EvalName.
       character*70     Name
       common /global/  Name
 
-      integer ctx
-      integer    ierror
+      integer(8)      ctx
+      integer         ierror
 
 
       call MPI_Init(ierror)
@@ -95,8 +95,8 @@ c     using all printable ASCII characters for the range.
 c
       subroutine N_InitString(ctx, p, pop) 
       include   'pgapackf.h'
-      integer ctx
-      integer    p, pop, i
+      integer(8)      ctx
+      integer         p, pop, i
     
       do i=PGAGetStringLength(ctx), 1, -1
          call PGASetCharacterAllele(ctx, p, pop, i,
@@ -116,7 +116,7 @@ c
       include          'pgapackf.h'
       character         Name(70)
       common /global/   Name
-      integer ctx
+      integer(8)        ctx
       integer           p1, p2, pop1, c1, c2, pop2
       integer           i, length
       character         a, b
@@ -153,7 +153,7 @@ c
       include          'pgapackf.h'
       character         Name(70)
       common /global/   Name
-      integer ctx
+      integer(8)        ctx
       integer           p1, pop1, p2, pop2
       integer           i, match
       character         a, b, c
@@ -187,7 +187,7 @@ c
       include          'pgapackf.h'
       character         Name(70)
       common /global/   Name
-      integer ctx
+      integer(8)        ctx
       integer           p, pop, i, count
       double precision  mr
 
@@ -214,7 +214,7 @@ c     we would print to the file "file".
 c     
       subroutine N_PrintString(ctx, file, p, pop)
       include          'pgapackf.h'
-      integer ctx
+      integer(8)        ctx
       integer           file, p, pop
       character         string(70)
 
@@ -235,7 +235,7 @@ c     the string yet.
 c
       integer function N_StopCond(ctx) 
       include          'pgapackf.h'
-      integer ctx
+      integer(8)        ctx
       integer           done, best
 
 
@@ -265,7 +265,7 @@ c     print string function to print the best string.
 c
       subroutine N_EndOfGeneration(ctx)
       include          'pgapackf.h'
-      integer ctx
+      integer(8)        ctx
       integer           best
 
       best = PGAGetBestIndex(ctx, PGA_NEWPOP)
@@ -286,7 +286,7 @@ c     the characters matching Name.
 c
       double precision function EvalName(ctx, p, pop)
       include          'pgapackf.h'
-      integer ctx
+      integer(8)        ctx
       integer           p, pop, i, count
       character         Name(70)
       common /global/   Name
